feat: upgrade Astro from v5 to v6
- Migrate content collections to Content Layer API with glob() loader
- Move src/content/config.ts to src/content.config.ts
- Replace entry.slug with entry.id across all components, pages, and utils
- Replace entry.render() with render(entry) from astro:content
- Upgrade astro-icon from v0.8 to v1.1 (new import paths + icon() integration)
- Replace @astrojs/tailwind with direct PostCSS tailwindcss plugin
- Upgrade @astrojs/mdx to v5, @astrojs/netlify to v7, @astrojs/react to v5
- Add custom stripIndexId function to maintain backward-compatible content IDs
- Temporarily disable astro-md-alternate (needs update for entry.body removal)
- Update all tests for slug→id migration and new API patterns
