+++ package.json
@@ -24,27 +24,29 @@
   },
   "dependencies": {
     "@astro-community/astro-embed-youtube": "^0.5.10",
-    "@astrojs/mdx": "^4.3.13",
-    "@astrojs/netlify": "^6.6.4",
-    "@astrojs/react": "^4.4.2",
-    "@astrojs/rss": "^4.0.15",
-    "@astrojs/sitemap": "^3.7.0",
-    "@astrojs/tailwind": "^6.0.2",
+    "@astrojs/mdx": "^5.0.0",
+    "@astrojs/netlify": "^7.0.1",
+    "@astrojs/react": "^5.0.0",
+    "@astrojs/rss": "^4.0.17",
+    "@astrojs/sitemap": "^3.7.1",
     "@atproto/api": "^0.19.0",
+    "@iconify-json/tabler": "^1.2.31",
     "@tailwindcss/forms": "^0.5.11",
     "@tailwindcss/typography": "^0.5.16",
+    "@testing-library/dom": "^10.4.1",
     "@types/animejs": "^3.1.13",
     "@types/date-fns": "^2.6.3",
     "@types/react": "^18.3.23",
     "@types/react-dom": "^18.3.7",
     "altcha-lib": "^1.4.1",
     "animejs": "^3.2.2",
-    "astro": "^5.18.0",
+    "astro": "^6.0.1",
     "astro-auto-import": "^0.5.1",
     "astro-embed": "^0.12.0",
-    "astro-icon": "0.8.2",
+    "astro-icon": "^1.1.5",
     "astro-md-alternate": "^0.1.0",
     "astro-seo": "^0.8.4",
+    "autoprefixer": "^10.4.27",
     "bluesky-comments": "^0.13.1",
     "csv-parse": "^6.1.0",
     "csv-stringify": "^6.6.0",
@@ -63,7 +65,7 @@
     "yaml": "^2.8.2"
   },
   "devDependencies": {
-    "@astrojs/check": "^0.9.6",
+    "@astrojs/check": "^0.9.7",
     "@fullhuman/postcss-purgecss": "^7.0.2",
     "@testing-library/jest-dom": "^6.6.4",
     "@testing-library/react": "^16.3.2",
