+++ src/components/About/TheStory.astro
@@ -4,7 +4,7 @@
  * Section 4 of the About page restructure
  * Previously ServicesSideImage, now restructured for the new narrative flow
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import ServiceCard from "@components/ServiceCard/ServiceCardSideImage.astro";
