+++ src/components/About/WhyItWorks.astro
@@ -4,7 +4,7 @@
  * Section 3 of the About page restructure
  */
 import Badge from "@components/Badge/Badge.astro";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 interface Approach {
   title: string;
