+++ src/components/Accordion/Accordion.astro
@@ -4,7 +4,7 @@
  * Uses the `name` attribute to enable "only one open at a time" behavior
  * when multiple accordions share the same group name.
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 interface Props {
   title: string;
