+++ src/components/Admonition/Admonition.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 interface Props {
   variant: "tip" | "caution" | "danger" | "info" | "terminal" | "prompt";
