+++ src/components/Author/AuthorBioBox.astro
@@ -22,7 +22,7 @@ const {
   credentials = [],
 } = author.data;
 
-const authorUrl = `/authors/${author.slug}`;
+const authorUrl = `/authors/${author.id}`;
 
 // Map social links with proper icon names (LinkedIn, GitHub, Bluesky, Website as per user requirement)
 const socialLinks = [
