+++ src/components/Blog/BlogNav.astro
@@ -15,7 +15,7 @@ const sortedPosts = posts.sort(
 );
 
 // Find current post index
-const currentIndex = sortedPosts.findIndex((p) => p.slug === currentSlug);
+const currentIndex = sortedPosts.findIndex((p) => p.id === currentSlug);
 const prevPost = currentIndex > 0 ? sortedPosts[currentIndex - 1] : null;
 const nextPost =
   currentIndex < sortedPosts.length - 1 ? sortedPosts[currentIndex + 1] : null;
@@ -53,7 +53,7 @@ const nextPost =
     {
       prevPost && (
         <a
-          href={`/post/${prevPost.slug}/`}
+          href={`/post/${prevPost.id}/`}
           class="group flex-1 rounded-lg border border-base-200 p-4 hover:border-primary-500 focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2 dark:border-base-800 dark:hover:border-primary-400"
           aria-label={`Previous article: ${prevPost.data.title}`}
           data-umami-event="nav-click"
@@ -84,7 +84,7 @@ const nextPost =
     {
       nextPost && (
         <a
-          href={`/post/${nextPost.slug}/`}
+          href={`/post/${nextPost.id}/`}
           class="group flex-1 rounded-lg border border-base-200 p-4 text-right hover:border-primary-500 focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2 dark:border-base-800 dark:hover:border-primary-400"
           aria-label={`Next article: ${nextPost.data.title}`}
           data-umami-event="nav-click"
