+++ src/components/Cta/CtaCardCenter.astro
@@ -3,7 +3,7 @@
  * * CTA card center with dark background and light text
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import Button from "@components/Button/Button.astro";
