+++ src/components/Cta/CtaCardSplit.astro
@@ -3,7 +3,7 @@
  * * CTA split with text left and button on the right, with primary color background and light text
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import Button from "@components/Button/Button.astro";
