+++ src/components/EnvironmentBanner/EnvironmentBanner.astro
@@ -1,7 +1,7 @@
 ---
 // src/components/EnvironmentBanner/EnvironmentBanner.astro
 import { CONFIG } from "src/config.ts";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 const isProduction = CONFIG.ALLOW_INDEXING;
 ---
