+++ src/components/Feature/FeatureToggleImage.astro
@@ -6,7 +6,7 @@
  * ! I copy all icons into the src/icons folder and recommend you do the same
  * ! If the icon is at src/icons/tabler/paint.svg then you put in "tabler/paint"
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { Image } from "astro:assets";
 
 // components
