+++ src/components/Feature/FeatureVideo.astro
@@ -4,7 +4,7 @@
  */
 
 import { Image } from "astro:assets";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import Badge from "@components/Badge/Badge.astro";
