+++ src/components/FeatureCard/FeatureCardMedium.astro
@@ -6,7 +6,7 @@
  * ! If the icon is at src/icons/tabler/paint.svg then you put in "tabler/paint"
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 interface Props {
   icon: string;
