+++ src/components/FeatureCard/FeatureCardSideImage.astro
@@ -4,7 +4,7 @@
  */
 
 import { Image } from "astro:assets";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { getRelativeLocaleUrl } from "astro:i18n";
 
 // components
