+++ src/components/Hero/HeroSideImage.astro
@@ -4,7 +4,7 @@
  *
  * There is a commented out "awards" section below in case you want that
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { Image } from "astro:assets";
 import SocialLinks from "@components/SocialLinks/SocialLinks.astro";
