+++ src/components/Nav/NavDropdown/MegaMenuDropdownToggle.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import NavLink from "../NavLink.astro";
