+++ src/components/Nav/NavDropdown/NavDropdown.astro
@@ -3,7 +3,7 @@
  * * Dropdown menu that opens on hover
  * This is for non-mobile devices
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import NavLink from "@components/Nav/NavLink.astro";
