+++ src/components/Nav/NavLink.astro
@@ -1,6 +1,6 @@
 ---
 import { getRelativeLocaleUrl } from "astro:i18n";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // utils
 import { getLocaleFromUrl } from "@js/localeUtils";
