+++ src/components/Newsletter/NewsletterForm.astro
@@ -17,7 +17,7 @@
  * - ALTCHA proof-of-work spam protection (privacy-friendly)
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import Button from "@components/Button/Button.astro";
 
 const altchaChallengeUrl = "/.netlify/functions/altcha-challenge";
