+++ src/components/Presentations/PresentationCard.astro
@@ -12,7 +12,7 @@ const { title, duration, intendedAudience, isWorkshop, image } =
 ---
 
 <article class="card group">
-  <a href={`/presentations/${presentation.slug}/`}>
+  <a href={`/presentations/${presentation.id}/`}>
     {
       image && (
         <Image
@@ -21,7 +21,7 @@ const { title, duration, intendedAudience, isWorkshop, image } =
           width={1600}
           height={900}
           class="w-full rounded-t-xl object-cover"
-          transition:name={`presentation-image-${presentation.slug}`}
+          transition:name={`presentation-image-${presentation.id}`}
           style="background: transparent;"
         />
       )
@@ -49,7 +49,7 @@ const { title, duration, intendedAudience, isWorkshop, image } =
     <!-- Title -->
     <h2 class="card-title">
       <a
-        href={`/presentations/${presentation.slug}/`}
+        href={`/presentations/${presentation.id}/`}
         class="hover:text-primary-500 dark:hover:text-primary-400"
       >
         {title}
@@ -68,7 +68,7 @@ const { title, duration, intendedAudience, isWorkshop, image } =
     <!-- Read More Button -->
     <div class="mt-6">
       <a
-        href={`/presentations/${presentation.slug}/`}
+        href={`/presentations/${presentation.id}/`}
         class="inline-flex items-center gap-2"
       >
         View Presentation
