+++ src/components/Projects/ProjectCard.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { Image } from "astro:assets";
 import Button from "@components/Button/Button.astro";
