+++ src/components/ServiceCard/ServiceCardIcon.astro
@@ -8,7 +8,7 @@
  * ! If the icon is at src/icons/tabler/paint.svg then you put in "tabler/paint"
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import Button from "@components/Button/Button.astro";
