+++ src/components/ServiceCard/ServiceCardSideImage.astro
@@ -5,7 +5,7 @@
  */
 
 import { Image } from "astro:assets";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { getRelativeLocaleUrl } from "astro:i18n";
 
 // components
