+++ src/components/Services/ServicesSideImage.astro
@@ -3,7 +3,7 @@
  * * These are alternating side cards with titles, text, and images
  * Recommend using 2 or 3 cards in a row
  */
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // components
 import ServiceCard from "@components/ServiceCard/ServiceCardSideImage.astro";
