+++ src/components/SocialCard/SocialCard.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 /**
  * Props for the SocialCard component
