+++ src/components/Testimonials/ColleagueTestimonialCard.astro
@@ -3,7 +3,7 @@
  * * Individual colleague testimonial card component
  */
 
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import type { ColleagueTestimonial } from "@config/colleagueTestimonialData";
 
 const { name, role, company, workRelationship, quote } =
