+++ src/components/Testimonials/TestimonialItemQuotes.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from "astro:assets";
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 
 // utils
 import { type TestimonialItem } from "@config/types/configDataTypes";
