+++ src/layouts/__tests__/BaseLayout.test.ts
@@ -18,7 +18,7 @@ describe("BaseLayout", () => {
           <meta name="description" content="${props.description}">
           <meta name="viewport" content="width=device-width, initial-scale=1">
           <meta charset="UTF-8">
-          <ViewTransitions />
+          <ClientRouter />
         </head>
         <body
           id="body"
