+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import { Icon } from "astro-icon";
+import { Icon } from "astro-icon/components";
 import { Image } from "astro:assets";
 import { getRelativeLocaleUrl } from "astro:i18n";
