+++ src/pages/[...page].astro
@@ -1,5 +1,5 @@
 ---
-import { type CollectionEntry, getCollection } from "astro:content";
+import { type CollectionEntry, getCollection, render } from "astro:content";
 
 // layout
 import BaseLayout from "@layouts/BaseLayout.astro";
@@ -18,7 +18,7 @@ export async function getStaticPaths() {
 
   // Note: otherPages don't have language-specific versions, so no need to filter by locale
   return otherPages.map((page) => ({
-    params: { page: page.slug },
+    params: { page: page.id },
     props: page,
   }));
 }
@@ -35,7 +35,7 @@ if (!page?.data) {
 }
 
 const { title, description } = page.data;
-const { Content } = await page.render();
+const { Content } = await render(page);
 ---
 
 <BaseLayout title={title} description={description}>
